Creacion y edicion del header
