+++ src/app/globals.css
@@ -31,3 +31,7 @@ body {
     text-wrap: balance;
   }
 }
+
+main {
+  background-color: #202020;
+}
