+++ src/app/ui/Header.js
@@ -1,0 +1,6 @@
+
+export default function Header () {
+    return (
+        <p>Hola</p>
+    );
+}
